PhysioTherapy Puzzle demonstration bug fix.
Bug fixed the puzzle to not allow new prefabs (blocks) to be spawned until the player has dropped chosen to drop the block. While it should not have happened during the demonstration, in all likelyhood when the scene was reloaded a block may be have been left behind long enough for the script to trigger a collision, thus result in new blocks spawning.